+++ src/main/java/com/alon/spring/crud/model/BaseEntity.java
@@ -3,6 +3,5 @@ package com.alon.spring.crud.model;
 import java.io.Serializable;
 
 public interface BaseEntity<T> extends Serializable {
-	public T getId();
-	public void setId(T id);
+    public T id();
 }