+++ src/main/java/com/alon/spring/crud/repository/specification/predicate/StartsWithPredicateBuilder.java
@@ -4,26 +4,26 @@ import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.Path;
 import javax.persistence.criteria.Predicate;
 
-public class GreaterThanPredicateBuilder implements PredicateBuilder {
-    
+public class StartsWithPredicateBuilder implements PredicateBuilder {
+
     private static PredicateBuilder INSTANCE;
 
     @Override
     public Predicate build(CriteriaBuilder criteriaBuilder, Path path, String value) {
-        
-        Comparable convertedValue = this.convertValue(path, value);
-        
-        return criteriaBuilder.greaterThan(path, convertedValue);
-        
+
+        value = String.format("%s%%", this.convertValue(path, value));
+
+        return criteriaBuilder.like(path, value);
+
     }
-    
+
     public static PredicateBuilder getInstance() {
-        
+
         if (INSTANCE == null)
-            INSTANCE = new GreaterThanPredicateBuilder();
-        
+            INSTANCE = new StartsWithPredicateBuilder();
+
         return INSTANCE;
-        
+
     }
-    
+
 }