+++ src/main/java/com/alon/spring/crud/service/CrudService.java
@@ -1,69 +1,68 @@
 package com.alon.spring.crud.service;
 
 import com.alon.querydecoder.Expression;
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
+import com.alon.querydecoder.SingleExpression;
 import com.alon.spring.crud.model.BaseEntity;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
-import javax.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Order;
 import org.springframework.data.jpa.domain.Specification;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface CrudService<ID, E extends BaseEntity<ID>, R extends JpaRepository<E, ID> & JpaSpecificationExecutor<E>> {
+import javax.validation.Valid;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+
+public interface CrudService<R extends JpaRepository & JpaSpecificationExecutor> {
     
     public R getRepository();
-
-    public List<Sort.Order> getDefaultSort();
     
-    default Page<E> list() {
+    default <E extends BaseEntity> Page<E> list() {
         return this.list(0, Integer.MAX_VALUE);
     }
     
-    default Page<E> list(Expression order) {
+    default <E extends BaseEntity> Page<E> list(SingleExpression order) {
         return this.list(0, Integer.MAX_VALUE, order);
     }
     
-    default Page<E> list(int page, int size) {
-        return this.list(page, size, null);
+    default <E extends BaseEntity> Page<E> list(int page, int size) {
+        return this.getRepository().findAll(PageRequest.of(page, size));
+    }
+
+    default <E extends BaseEntity> Page<E> list(int page, int size, SingleExpression order) {
+        return this.getRepository().findAll(Hidden.buildPageable(page, size, order));
     }
     
-    default Page<E> list(Specification<E> specification, int page, int size) {
-        return this.list(specification, page, size, null);
+    default <E extends BaseEntity> Page<E> list(Specification<E> specification, int page, int size) {
+        return this.getRepository().findAll(specification, PageRequest.of(page, size));
     }
 
-    default Page<E> list(int page, int size, Expression order) {
-        return this.getRepository().findAll(Hidden.buildPageable(page, size, order, this.getDefaultSort()));
+    default <E extends BaseEntity> Page<E> list(Specification<E> specification, int page, int size, SingleExpression order) {
+        return this.getRepository().findAll(specification, Hidden.buildPageable(page, size, order));
     }
 
-    default Page<E> list(Specification<E> specification, int page, int size, Expression order) {
-        return this.getRepository().findAll(specification, Hidden.buildPageable(page, size, order, this.getDefaultSort()));
-    }
-
-    default E create(@Valid E entity) throws CreateException {
+    default <E extends BaseEntity> E create(@Valid E entity) throws CreateException {
         try {
             entity = Hidden.executeHook(this, entity, LifeCycleHook.BEFORE_CREATE);
-            entity = this.getRepository().save(entity);
+            entity = (E) this.getRepository().save(entity);
             return Hidden.executeHook(this, entity, LifeCycleHook.AFTER_CREATE);
         } catch (Throwable ex) {
             throw new CreateException(ex.getMessage(), ex);
         }
     }
 
-    default E read(ID id) {
-        return this.getRepository().findById(id).get();
+    default <ID, E extends BaseEntity<ID>> E read(ID id) {
+        return (E) this.getRepository().findById(id).get();
     }
 
-    default E update(@Valid E entity) throws UpdateException {
-        if (entity.getId() == null)
+    default <E extends BaseEntity> E update(@Valid E entity) throws UpdateException {
+        if (entity.id() == null)
             throw new UpdateException("Unmanaged entity. Use the create method.");
             
         try {
@@ -75,7 +74,7 @@ public interface CrudService<ID, E extends BaseEntity<ID>, R extends JpaReposito
         }
     }
 
-    default void delete(ID id) throws DeleteException {
+    default <ID> void delete(ID id) throws DeleteException {
         try {
             Hidden.executeHook(this, id, LifeCycleHook.BEFORE_DELETE);
             this.getRepository().deleteById(id);
@@ -85,32 +84,32 @@ public interface CrudService<ID, E extends BaseEntity<ID>, R extends JpaReposito
         }
     }
     
-    default CrudService addBeforeCreateHook(CheckedFunction<E, E> function) {
+    default <E extends BaseEntity> CrudService addBeforeCreateHook(Function<E, E> function) {
     	Hidden.getServiceHooks(this).get(LifeCycleHook.BEFORE_CREATE).add(function);
     	return this;
     }
     
-    default CrudService addAfterCreateHook(CheckedFunction<E, E> function) {
+    default <E extends BaseEntity> CrudService addAfterCreateHook(Function<E, E> function) {
     	Hidden.getServiceHooks(this).get(LifeCycleHook.AFTER_CREATE).add(function);    	
     	return this;
     }
     
-    default CrudService addBeforeUpdateHook(CheckedFunction<E, E> function) {
+    default <E extends BaseEntity> CrudService addBeforeUpdateHook(Function<E, E> function) {
     	Hidden.getServiceHooks(this).get(LifeCycleHook.BEFORE_UPDATE).add(function);    	
     	return this;
     }
     
-    default CrudService addAfterUpdateHook(CheckedFunction<E, E> function) {
+    default <E extends BaseEntity> CrudService addAfterUpdateHook(Function<E, E> function) {
     	Hidden.getServiceHooks(this).get(LifeCycleHook.AFTER_UPDATE).add(function);    	
     	return this;
     }
     
-    default CrudService addBeforeDeleteHook(CheckedFunction<Long, Long> function) {
+    default <E extends BaseEntity> CrudService addBeforeDeleteHook(Function<Long, Long> function) {
     	Hidden.getServiceHooks(this).get(LifeCycleHook.BEFORE_DELETE).add(function);    	
     	return this;
     }
     
-    default CrudService addAfterDeleteHook(CheckedFunction<Long, Long> function) {
+    default <E extends BaseEntity> CrudService addAfterDeleteHook(Function<Long, Long> function) {
     	Hidden.getServiceHooks(this).get(LifeCycleHook.AFTER_DELETE).add(function);    	
     	return this;
     }
@@ -125,22 +124,19 @@ public interface CrudService<ID, E extends BaseEntity<ID>, R extends JpaReposito
     }
     
     /**
-     * This class has the function of simulating private methods, which is not 
-     * allowed in interfaces
+     * This class has the function of restricting methods, 
+     * because private methods are not allowed in interfaces.
      */
     class Hidden {
         
-        private static Map<CrudService, Map<LifeCycleHook, List<CheckedFunction>>> GLOBAL_HOOKS = new HashMap<>();
+        private static Map<CrudService, Map<LifeCycleHook, List<Function>>> GLOBAL_HOOKS = new HashMap<>();
         
-        private static Pageable buildPageable(int page, int size, Expression orders, List<Order> defaultSort) {
-            return PageRequest.of(page, size, buildSort(orders, defaultSort));
+        private static Pageable buildPageable(int page, int size, SingleExpression orders) {
+            return PageRequest.of(page, size, buildSort(orders));
         }
 
-        private static Sort buildSort(Expression order, List<Order> defaultSort) {
+        private static Sort buildSort(SingleExpression order) {
             
-            if (order == null || order.getField() == null)
-                return Sort.by(defaultSort);
-
             List<Order> orders = new ArrayList<>();
 
             do {
@@ -151,7 +147,7 @@ public interface CrudService<ID, E extends BaseEntity<ID>, R extends JpaReposito
                 else
                     orders.add(Order.asc(order.getField()));
 
-                order = (Expression) order.getNext();
+                order = (SingleExpression) order.getNext();
             } while (order != null);
 
             return Sort.by(orders);
@@ -160,22 +156,22 @@ public interface CrudService<ID, E extends BaseEntity<ID>, R extends JpaReposito
     
         private static <S extends CrudService, P> P executeHook(S service, P param, LifeCycleHook hookType) throws Throwable {
             
-            List<CheckedFunction> hooks = getHook(service, hookType);
+            List<Function> hooks = getHook(service, hookType);
 
-            for (CheckedFunction hook : hooks)
+            for (Function hook : hooks)
                 param = (P) hook.apply(param);
 
             return param;
             
         }
 
-        private static <T extends CrudService> List<CheckedFunction> getHook(T service, LifeCycleHook hookType) {
+        private static <T extends CrudService> List<Function> getHook(T service, LifeCycleHook hookType) {
             return getServiceHooks(service).get(hookType);
         }
         
-        private static <T extends CrudService> Map<LifeCycleHook, List<CheckedFunction>> getServiceHooks(T service) {
+        private static <T extends CrudService> Map<LifeCycleHook, List<Function>> getServiceHooks(T service) {
             
-            Map<LifeCycleHook, List<CheckedFunction>> hooks = GLOBAL_HOOKS.get(service);
+            Map<LifeCycleHook, List<Function>> hooks = GLOBAL_HOOKS.get(service);
 
             if (hooks == null)
                 hooks = initHooks(service);
@@ -184,9 +180,9 @@ public interface CrudService<ID, E extends BaseEntity<ID>, R extends JpaReposito
             
         }
         
-        private static <T extends CrudService> Map<LifeCycleHook, List<CheckedFunction>> initHooks(T service) {
+        private static <T extends CrudService> Map<LifeCycleHook, List<Function>> initHooks(T service) {
             
-            Map<LifeCycleHook, List<CheckedFunction>> hooks = new HashMap<>();
+            Map<LifeCycleHook, List<Function>> hooks = new HashMap<>();
 
             for (LifeCycleHook hook : LifeCycleHook.values())
                 hooks.put(hook, new ArrayList<>());